extract device actions into separate mixin
