+++ frontend/src/mixins/DeviceActions.js
@@ -1,0 +1,192 @@
+import { mapState } from 'vuex'
+
+import ApplicationApi from '../api/application.js'
+
+import deviceApi from '../api/devices.js'
+import instanceApi from '../api/instances.js'
+import teamApi from '../api/team.js'
+import Alerts from '../services/alerts.js'
+import Dialog from '../services/dialog.js'
+
+export default {
+    computed: {
+        ...mapState('account', ['team']),
+        displayingTeam () {
+            return this.team !== null && !this.displayingInstance && !this.displayingApplication
+        },
+        displayingApplication () {
+            return Object.hasOwnProperty.call(this, 'application') && this.application !== null && !this.displayingInstance
+        },
+        displayingInstance () {
+            return Object.hasOwnProperty.call(this, 'instance') && this.instance !== null
+        },
+        teamDeviceCount () {
+            return this.team.deviceCount + this.deviceCountDeltaSincePageLoad
+        }
+    },
+    data () {
+        return {
+            allDeviceStatuses: new Map(), // every device known
+            deviceCountDeltaSincePageLoad: 0
+        }
+    },
+    emits: ['delete-device'],
+    methods: {
+        deviceAction (action, deviceId) {
+            const device = this.devices.get(deviceId)
+            if (action === 'edit') {
+                this.showEditDeviceDialog(device)
+            } else if (action === 'delete') {
+                Dialog.show({
+                    header: 'Delete Device',
+                    kind: 'danger',
+                    text: 'Are you sure you want to delete this device? Once deleted, there is no going back.',
+                    confirmLabel: 'Delete'
+                }, async () => {
+                    try {
+                        await deviceApi.deleteDevice(device.id)
+                        Alerts.emit('Successfully deleted the device', 'confirmation')
+                        this.deleteLocalCopyOfDevice(device)
+                        this.$emit('delete-device')
+                    } catch (err) {
+                        Alerts.emit('Failed to delete device: ' + err.toString(), 'warning', 7500)
+                    }
+                })
+            } else if (action === 'updateCredentials') {
+                this.$refs.deviceCredentialsDialog.show(device)
+            } else if (action === 'removeFromProject') {
+                Dialog.show({
+                    header: 'Remove Device from Instance',
+                    kind: 'danger',
+                    text: 'Are you sure you want to remove this device from the instance? This will stop the flows running on the device.',
+                    confirmLabel: 'Remove'
+                }, async () => {
+                    await deviceApi.updateDevice(device.id, { instance: null })
+
+                    if (this.displayingInstance) {
+                        this.deleteLocalCopyOfDevice(device)
+                    } else {
+                        this.updateLocalCopyOfDevice({ ...device, instance: undefined, application: undefined, ownerType: '' })
+                    }
+
+                    Alerts.emit('Successfully removed the device from the instance.', 'confirmation')
+                })
+            } else if (action === 'removeFromApplication') {
+                Dialog.show({
+                    header: 'Remove Device from Application',
+                    kind: 'danger',
+                    text: 'Are you sure you want to remove this device from the application? This will stop the flows running on the device.',
+                    confirmLabel: 'Remove'
+                }, async () => {
+                    await deviceApi.updateDevice(device.id, { application: null })
+
+                    if (this.displayingApplication) {
+                        this.deleteLocalCopyOfDevice(device)
+                    } else {
+                        this.updateLocalCopyOfDevice({ ...device, instance: undefined, application: undefined, ownerType: '' })
+                    }
+
+                    Alerts.emit('Successfully removed the device from the application.', 'confirmation')
+                })
+            } else if (action === 'assignToProject') {
+                this.$refs.deviceAssignInstanceDialog.show(device)
+            } else if (action === 'assignToApplication') {
+                this.$refs.deviceAssignApplicationDialog.show(device, false)
+            }
+        },
+
+        showEditDeviceDialog (device) {
+            this.$refs.teamDeviceCreateDialog.show(device)
+        },
+
+        updateLocalCopyOfDevice (device) {
+            if (!this.allDeviceStatuses.get(device.id)) {
+                this.deviceCountDeltaSincePageLoad++
+            }
+
+            // Only grab status props to avoid polluting allDeviceStatuses with extra info
+            const currentDeviceStatus = this.allDeviceStatuses.get(device.id)
+            if (currentDeviceStatus) {
+                const updatedDeviceStatusPropsOnly = Object.keys(currentDeviceStatus).reduce((acc, key) => {
+                    acc[key] = device[key]
+                    return acc
+                }, { ...currentDeviceStatus })
+                this.allDeviceStatuses.set(device.id, updatedDeviceStatusPropsOnly)
+            } else {
+                this.allDeviceStatuses.set(device.id, device)
+            }
+
+            this.devices.set(device.id, device)
+        },
+
+        deviceCreated (device) {
+            if (device) {
+                setTimeout(() => {
+                    this.$refs.deviceCredentialsDialog.show(device)
+                }, 500)
+
+                this.updateLocalCopyOfDevice(device)
+            }
+        },
+
+        deviceUpdated (device) {
+            this.updateLocalCopyOfDevice(device)
+        },
+
+        deleteLocalCopyOfDevice (device) {
+            if (this.allDeviceStatuses.get(device.id)) {
+                this.deviceCountDeltaSincePageLoad--
+            }
+            this.allDeviceStatuses.delete(device.id)
+            this.devices.delete(device.id)
+
+            if ( // updates the application device count if present
+                Object.hasOwnProperty.call(this, 'application') &&
+                this.application !== null &&
+                Object.hasOwnProperty.call(this.application, 'deviceCount') &&
+                !isNaN(this.application.deviceCount) &&
+                this.application.deviceCount >= 1
+            ) {
+                this.application.deviceCount--
+            }
+        },
+
+        // Actual fetching methods
+        async fetchData (nextCursor = null, limit = null, extraParams = { statusOnly: false }) {
+            const query = null // handled via extraParams
+            if (this.displayingInstance) {
+                return await instanceApi.getInstanceDevices(this.instance.id, nextCursor, limit, query, extraParams)
+            }
+
+            if (this.displayingApplication) {
+                return await ApplicationApi.getApplicationDevices(this.application.id, nextCursor, limit, query, extraParams)
+            }
+
+            if (this.displayingTeam) {
+                return await teamApi.getTeamDevices(this.team.id, nextCursor, limit, query, extraParams)
+            }
+
+            console.warn('Trying to fetch data without a loaded model.')
+
+            return null
+        },
+
+        async fetchAllDeviceStatuses (reset = false) {
+            const data = await this.fetchData(null, null, { statusOnly: true })
+
+            if (reset) {
+                this.allDeviceStatuses = new Map()
+            }
+
+            if (data.meta?.next_cursor || data.devices.length < data.count) {
+                console.warn('Device Status API should not be paginating')
+            }
+
+            data.devices.forEach(device => {
+                this.allDeviceStatuses.set(device.id, device)
+            })
+
+            this.loadingStatuses = false
+        }
+    }
+}
